ask-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
qdrant-client==1.12.0
sentence-transformers==3.2.1
openai==1.54.3
boto3==1.35.36
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
llama-cloud-services==0.1.0
